requirements.txt
spotipy
lyricsgenius
moviepy
mutagen
youtubesearchpython
tenacity
pytube
deezer-python
python-dotenv
requests
